// src/advert/schema/advert.dto.ts
export class AdvertDTO {
  caption: string;
  message: string;
  email: string;
  duration: string;
  link: string;
  action: string;
  audience: string;
  image: string;
  shares: number;
  likes: number;
}


export class CreateAdvertDTO {
  caption: string;
  message: string;
  email: string;
  duration: string;
  link: string;
  action: string;
  audience: string;
  imageFile: string;
}